make table component.
